assaddsa

+++ admin/tampil_status.php
@@ -3,9 +3,11 @@
 
   if($_POST["query"] != 0)
   {
+   $a = explode(",",$_POST["q"]);
+   $s = "'". implode("', '", $a) ."'";
    $search_array = explode(",", $_POST["query"]);
    $search_text = "'" . implode("', '", $search_array) . "'";
-   $query = "SELECT * FROM hasil_tes WHERE status IN (".$search_text.") ORDER BY id_hasil_tes DESC";
+   $query = "SELECT * FROM hasil_tes WHERE status IN (".$search_text.") AND tahun IN (".$s.") ORDER BY status DESC";
   }
   else
   {
@@ -42,7 +44,7 @@
      <center>
      <a href="lihat-tes.php?id_hasil_tes='.$row["id_hasil_tes"].'"><i class="fa fa-eye"></i></a>
      <a href="form-edit-hasil.php?id_hasil_tes='.$row["id_hasil_tes"].'"><i class="fa fa-edit"></i></a>
-     <a href="hapus-hasil.php?id_hasil_tes='.$row["id_hasil_tes"].'"><i class="fa fa-trash"></i></a>
+     <a href="hapus-hasil.php?id_hasil_tes='.$row["id_hasil_tes"].'" onclick = "return confirm(/Yakin Ingin hapus data ini/)"><i class="fa fa-trash"></i></a>
      <a href="print-semua-hasil.php?id_hasil_tes='.$row['id_hasil_tes'].'"><i class="fa fa-print"></i></a>
      <center>
      </td>
